+++ shared/src/commonMain/kotlin/com/lwtor/xhunter/di/SharedModule.kt
@@ -1,0 +1,12 @@
+package com.lwtor.xhunter.di
+
+import com.arkivanov.decompose.ComponentContext
+import com.lwtor.xhunter.ui.main.DefaultRootComponent
+import com.lwtor.xhunter.ui.main.RootComponent
+import org.koin.dsl.module
+
+val sharedModule = module {
+    factory<RootComponent> { (componentContext: ComponentContext) ->
+        DefaultRootComponent(componentContext)
+    }
+}